refactor: shared/ui dialog 분리
