+++ src/shared/ui/dialog/index.ts
@@ -1,0 +1,4 @@
+export * from "./Dialog";
+export * from "./DialogContent";
+export * from "./DialogHeader";
+export * from "./DialogTitle"
